+++ src/components/Login/Login.css
@@ -1,6 +1,6 @@
 @font-face {
 	font-family: "Cubano";
-	src: url(../assets/fonts/Cubano/Cubano.ttf);
+	src: url(../../assets/fonts/Cubano/Cubano.ttf);
 }
 
 .Login {
