+++ src/components/Player/Player.jsx
@@ -1,6 +1,6 @@
-import Footer from "./Footer";
-import Sidebar from "./Sidebar";
-import Body from "./Body";
+import Footer from "../Footer/Footer";
+import Sidebar from "../Sidebar/Sidebar";
+import Body from "../Body/Body";
 import "./Player.css";
 
 const Player = ({ spotify }) => {
